--- src/containers/About/index.tsx
@@ -1,12 +1,0 @@
-import { FunctionComponent } from "react";
-
-const About: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>About</h1>
-            
-        </div>
-    )
-}
-
-export default About;
