--- src/containers/Feed/index.tsx
@@ -1,11 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Feed: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>Feed</h1>
-        </div>
-    )
-}
-
-export default Feed;
