--- src/containers/Home/index.tsx
@@ -1,11 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Home: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>Home</h1>
-        </div>
-    )
-}
-
-export default Home;
