--- src/containers/Login/index.tsx
@@ -1,11 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Login: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>Login</h1>
-        </div>
-    )
-}
-
-export default Login;
