--- src/containers/Notification/index.tsx
@@ -1,11 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Notification: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>Notification</h1>
-        </div>
-    )
-}
-
-export default Notification;
