+++ src/containers/Pages/about.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const About: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>About</h1>
+    </div>
+  );
+};
+
+export default About;
