+++ src/containers/Pages/feed.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const Feed: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>Feed </h1>
+    </div>
+  );
+};
+
+export default Feed;
