+++ src/containers/Pages/home.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const Home: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>Home</h1>
+    </div>
+  );
+};
+
+export default Home;
