+++ src/containers/Pages/login.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const Login: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>Login</h1>
+    </div>
+  );
+};
+
+export default Login;
