+++ src/containers/Pages/notification.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const Notification: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>Notification</h1>
+    </div>
+  );
+};
+
+export default Notification;
