+++ src/containers/Pages/profile.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const Profile: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>Profile</h1>
+    </div>
+  );
+};
+
+export default Profile;
