+++ src/containers/Pages/signup.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from "react";
+
+const Signup: FunctionComponent = () => {
+  return (
+    <div>
+      <h1>Signup</h1>
+    </div>
+  );
+};
+
+export default Signup;
