--- src/containers/Profile/index.tsx
@@ -1,11 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Profile: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>Profile</h1>
-        </div>
-    )
-}
-
-export default Profile;
