--- src/containers/Signup/index.tsx
@@ -1,11 +1,0 @@
-import { FunctionComponent } from "react";
-
-const Signup: FunctionComponent = () => {
-    return (
-        <div>
-            <h1>Signup</h1>
-        </div>
-    )
-}
-
-export default Signup;
