+++ src/reportWebVitals.tsx
@@ -1,16 +1,16 @@
-import { ReportHandler } from 'web-vitals';
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+import { ReportHandler } from "web-vitals";
 
 const reportWebVitals = (onPerfEntry?: ReportHandler) => {
   if (onPerfEntry && onPerfEntry instanceof Function) {
-    import('web-vitals').then(({ getCLS, getFID, getFCP, getLCP, getTTFB }) => {
+    import("web-vitals").then(({ getCLS, getFID, getFCP, getLCP, getTTFB }) => {
       getCLS(onPerfEntry);
       getFID(onPerfEntry);
       getFCP(onPerfEntry);
       getLCP(onPerfEntry);
       getTTFB(onPerfEntry);
-    }); 
+    });
   }
 };
 
 export default reportWebVitals;
-
